fix: Remove fatal errors outside main.go to improve flexibility
Ran into an issue where I could not grant all the permissions needed for
the exporter and only really needed a subset of the metrics. The
exporter always attempts to fetch most of the metrics from CF. For some
of these the fetching functions would fatally exit the program. This
seems a bit aggressive. The choice was either allow the error and carry
on or add more flags/options to control which metrics are actually
fetched. Removing the fatal errors seemed like the better option.

- Remove all fatal errors outside main.go (in cloudflare.go)
- Errors now log followed by `nil` return values for slices/maps which
  calling code handles